Extract FlowMeterSection and tighten method formatting
